solution: Fixed Task Point for theme 'oop-classes-and-objects'

+++ src/main/java/com/rakovets/course/java/core/practice/oop/classesandobjects/Point.java
@@ -1,33 +1,33 @@
 package com.rakovets.course.java.core.practice.oop.classesandobjects;
 
 public class Point {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int distance(Point newPoint) {
+    public double distance(Point newPoint) {
         double distance;
         distance = Math.sqrt((newPoint.x - this.x) * (newPoint.x - this.x) + (newPoint.y - this.y) * (newPoint.y - this.y));
-        return (int) distance;
+        return distance;
     }
 }
